backend/requirements.txt
flask
bcrypt
sqlite3
python-dotenv